refactor(auth): move UserProfile type to common types and resolve deps
